fix: some warnings

+++ src/base/KgDct.cpp
@@ -23,7 +23,7 @@ void KgDct::forward(const double* in, double* out) const
 	// TODO: 使用矩阵运算
 	for(unsigned k = 0; k < odim(); k++) {
 		out[k] = 0;
-		for(size_t j = 0; j < idim(); j++)
+		for(unsigned j = 0; j < idim(); j++)
 			out[k] += in[j] * (*this)(k, j);
 	}
 }
@@ -33,9 +33,9 @@ void KgDct::backward(const double* in, double* out) const
 {
 	double factor = 2.0 / idim(); // TODO: 
 
-	for(size_t j = 0; j < idim(); j++) {
+	for(unsigned j = 0; j < idim(); j++) {
 		out[j] = 0.5 * in[0] * (*this)(0, j);
-		for(size_t k = 1; k < odim(); k++)
+		for(unsigned k = 1; k < odim(); k++)
 			out[j] += in[k] * (*this)(k, j);
 		out[j] *= factor;
 	}
@@ -47,18 +47,18 @@ void KgDct::initCosinTable_()
 	//generate cosin table
 	cosinTable_ = new double[idim() * odim()]; // if dim is too big, this will crash
 
-	for (size_t k = 0; k < odim(); k++)
-		for (size_t j = 0; j < idim(); j++)
+	for (unsigned k = 0; k < odim(); k++)
+		for (unsigned j = 0; j < idim(); j++)
 			(*this)(k, j) = cos(KtuMath<double>::pi * k * (j + 0.5) / idim());
 
 	if (norm_) {
 		double norm = sqrt(1.0 / idim());  // normalizer for X_0.
-		for (size_t j = 0; j < idim(); j++)
+		for (unsigned j = 0; j < idim(); j++)
 			(*this)(0, j) *= norm;  // TODO: 其实(0, j) == 1
 
 		norm = sqrt(2.0 / idim());  // normalizer for other elements.
-		for (size_t k = 1; k < odim(); k++)
-			for (size_t j = 0; j < idim(); j++)
+		for (unsigned k = 1; k < odim(); k++)
+			for (unsigned j = 0; j < idim(); j++)
 				(*this)(k, j) *= norm;
 	}
 }
